FIRST INTEGRATED LOOPER AND TEENSY-EXPRESSION TEST Turned on "WITH VOLUMES" to no apparent ill-effect.
uiWindow is now responding (per updateFrame)
to the serial port and directly calling looper commands
or setting controller volumes (bypassing my whole "midi
subsystem")

Also note that the loop volume pedal is sluggish.

A major code cleanup is demanded after a proof of concept
and playing with it, finally, for a little while.